// MeskChatClient/src/types/message.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * MeskChatApplication.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { User } from './user';

export interface Message {
  senderId?: string;
  sender?: User;
  receiverId?: string;
  receiver?: User;
  text?: string;
  sendAt?: string;
  isRead?: boolean;
  /** @nullable */
  readAt?: string | null;
  id?: string;
  isDeleted?: boolean;
  createdAt?: string;
  /** @nullable */
  updatedAt?: string | null;
  /** @nullable */
  deletedAt?: string | null;
}