// MeskChatClient/src/types/responseEntityOfListOfMessage.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * MeskChatApplication.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { Message } from './message';
import type { ResponseEntityOfListOfMessageValidationErrors } from './responseEntityOfListOfMessageValidationErrors';

export interface ResponseEntityOfListOfMessage {
  isSuccess?: boolean;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  data?: Message[] | null;
  statusCode?: number;
  /** @nullable */
  validationErrors?: ResponseEntityOfListOfMessageValidationErrors;
}